Move menu scripts under the same folder
