+++ Scripts/Interactables/Bonfire.cs
@@ -18,21 +18,21 @@ public partial class Bonfire : Interactable {
     SetStillFrame(GD.Load<Texture2D>("res://Assets/Sprites/Bonfire.png"));
   }
 
-  public override void OnNudge(Player player, Camera camera) {
-    OnInteract(player, camera);
-    base.OnNudge(player, camera);
+  public override void OnNudge(Player player, World world) {
+    OnInteract(player, world);
+    base.OnNudge(player, world);
   }
 
-  public override void OnInteract(Player player, Camera camera) {
+  public override void OnInteract(Player player, World world) {
     if (State == BonfireState.Extinguished) return;
 
-    base.OnInteract(player, camera);
+    base.OnInteract(player, world);
 
     if (State == BonfireState.Lit) {
       SpawnFloatingLabel("Rested", color: Global.White);
       var healAmount = player.MaxHealth - player.Health;
       if (healAmount > 0) player.Heal(healAmount);
-      player.InventoryMenu.Open();
+      world.InventoryMenu.Open();
       if (!Extinguishable) return;
       SetStillFrame(GD.Load<Texture2D>("res://Assets/Sprites/ExtinguishedBonfire.png"), "extinguished");
       State = BonfireState.Extinguished;
@@ -42,7 +42,7 @@ public partial class Bonfire : Interactable {
     State = BonfireState.Lit;
     Interacted = true;
 
-    camera.Shake(25);
+    world.Camera.Shake(25);
     SetAnimation("idle", GD.Load<Texture2D>("res://Assets/Animations/Bonfire.png"), 4, 12);
     SpawnFallingLabel("Lit!");
   }
