+++ Scripts/Interactables/Door.cs
@@ -13,14 +13,14 @@ public partial class Door : Interactable {
     base._Process(delta);
   }
 
-  public override void OnInteract(Player player, Camera camera) {
-    base.OnInteract(player, camera);
+  public override void OnInteract(Player player, World world) {
+    base.OnInteract(player, world);
 
     if (Interacted) return;
 
     Interacted = true;
     Solid = false;
-    camera.Shake(10);
+    world.Camera.Shake(10);
     SpawnFallingLabel("Opened!");
     Play("open");
   }
