+++ Scripts/Interactables/Ladder.cs
@@ -11,8 +11,8 @@ public partial class Ladder : Interactable {
     SetStillFrame(ResourceLoader.Load<Texture2D>("res://Assets/Sprites/Ladder.png"));
   }
 
-  public override void OnNudge(Player player, Camera camera) {
-    base.OnNudge(player, camera);
+  public override void OnNudge(Player player, World world) {
+    base.OnNudge(player, world);
 
     if (player.TurnsLived - _lastPlayerTurnsLived == 1) {
       Data.LoadLobbyData();
@@ -28,7 +28,7 @@ public partial class Ladder : Interactable {
 
     _lastPlayerTurnsLived = player.TurnsLived;
 
-    camera.Shake(10);
+    world.Camera.Shake(10);
     SpawnFloatingLabel("Leave dungeon?");
   }
 }
