+++ Scripts/Interactables/Workbench.cs
@@ -10,20 +10,20 @@ public partial class Workbench : Interactable {
     SetStillFrame(ResourceLoader.Load<Texture2D>("res://Assets/Sprites/Workbench.png"));
   }
 
-  public override void OnNudge(Player player, Camera camera) {
+  public override void OnNudge(Player player, World world) {
     if (Interacted) return;
 
-    base.OnNudge(player, camera);
+    base.OnNudge(player, world);
 
     Blink();
-    player.WorkbenchMenu.Open();
+    world.WorkbenchMenu.Open();
 
     Interacted = true;
 
     player.OnMoveEvent += OnPlayerMove;
     _player = player;
 
-    camera.Shake(10);
+    world.Camera.Shake(10);
     //Play("open");
   }
